refactor: Reorganizes model files to simplify maintenance.
